feat: Add script to extract device info from the project
The script needs to receive the project path as argument.
Furthermore, the project name should be identical to the
directory name, since the script expects to find a binary
file with the name build/${PROJECT-DIR}.bin

Furthermore, for the MAC to be extracted, the device should
be connected. Feel free to modify the script based on your needs.

PR: https://github.com/nubificus/esp32-ota-update/pull/6
Signed-off-by: Ilias Lagomatis <ilago@nubificus.co.uk>
Reviewed-by: Anastassios Nanos <ananos@nubificus.co.uk>
Approved-by: Anastassios Nanos <ananos@nubificus.co.uk>